vexternal1 models: float occurrences & successions

+++ HarmonyDB.Index/HarmonyDB.Index.Api.Model/VExternal1/StructureLoopsResponseDistributions.cs
@@ -3,8 +3,8 @@ namespace HarmonyDB.Index.Api.Model.VExternal1;
 public record StructureLoopsResponseDistributions
 {
     public required int[] TotalSongs { get; init; }
-    public required int[] TotalOccurrences { get; init; }
-    public required int[] TotalSuccessions { get; init; }
+    public required float[] TotalOccurrences { get; init; }
+    public required float[] TotalSuccessions { get; init; }
     public required float[] TonalityConfidence { get; init; }
     public required float[] TonicConfidence { get; init; }
     public required float[] TonicScore { get; init; }
